Change jetstream to use v8 api
Currently jetstream uses the v6 api and this changes it to use the v8 api instead. It also adds some tests for the api.
